+++ packages/failures/lib/src/failures/generic_failure.dart
@@ -3,6 +3,7 @@ import 'failure.dart';
 class GenericFailure extends Failure<Object> {
   GenericFailure(
     super.error, {
+    super.message,
     super.extra,
     super.stackTrace,
   });